Revert "Move trace op to pten (#39227)"
This reverts commit d7dddf9486bf7e39456be44b5a747db16644e818.
