--- paddle/pten/kernels/cpu/trace_grad_kernel.cc
@@ -1,31 +1,0 @@
-// Copyright (c) 2022 PaddlePaddle Authors. All Rights Reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#include "paddle/pten/kernels/trace_grad_kernel.h"
-#include "paddle/pten/kernels/impl/trace_kernel_impl.h"
-
-#include "paddle/pten/backends/cpu/cpu_context.h"
-#include "paddle/pten/core/kernel_registry.h"
-
-PT_REGISTER_KERNEL(trace_grad,
-                   CPU,
-                   ALL_LAYOUT,
-                   pten::TraceGradKernel,
-                   float,
-                   double,
-                   int,
-                   int64_t,
-                   paddle::platform::float16,
-                   paddle::platform::complex<float>,
-                   paddle::platform::complex<double>) {}
