--- paddle/pten/kernels/trace_grad_kernel.h
@@ -1,30 +1,0 @@
-// Copyright (c) 2022 PaddlePaddle Authors. All Rights Reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#pragma once
-
-#include "paddle/pten/core/dense_tensor.h"
-
-namespace pten {
-
-template <typename T, typename Context>
-void TraceGradKernel(const Context& ctx,
-                     const DenseTensor& out_grad,
-                     const DenseTensor& x,
-                     int offset,
-                     int axis1,
-                     int axis2,
-                     DenseTensor* in_grad);
-
-}  // namespace pten
